pyproject.toml: hardcode list of package
Commit e5e2d69c, which removed the setuptools 'find' config from
pyproject.toml, was not the correct fix for 'pip install': it causes pip
to become very confused and create an endless directory hierarchy under
build/lib/build/lib/... which gets deeper every time the package is
installed.

Instead, hardcode the list of packages to install in pyproject.toml.
This fixes the original problem (some files are not installed) as well
as the new problem.

I am not convinced this is the best fix for the issue in the long term,
but it makes the build functional for now.

Fixes:	e5e2d69caa49 ("pyproject: remove setuptools 'find' config")