status_detail: show a notice for opened sensitive statuses
When a sensitive status has been opened, continue to show a sensitive
status warning, but draw it in yellow rather than red.  The notice can
be removed in CSS for users who don't want it.

This allows the user to know which toots were sensitive even when using
the -S option.