status_list: show reblog status
Show an 'R' flag in status_list if a status was reblogged.

This changes the StatusListItem from a single Static into a horizontal
layout of three Labels, for the timestamp, flags and account.  A
side-effect of this is that the style of the labels (colours, etc.) can
now be customised via CSS.

The list item is correctly truncated (not wrapped) when the status list
view is too narrow to show the entire account name, but it doesn't add
ellipses when truncating.

While here, add an id tag for the main layout in app so we can query for
it explicitly, instead of querying for any Horizontal.