status_list: add 'width' to Labels
Without 'width: auto' (or 'width: 2' in the case of flags), the width of
the Labels can randomly change when scrolling up or down the list, which
breaks the display.  I'm not sure why this happens (it seems like it
might be a Textual bug), but specifying the width for the elements
appears to fix it.