support replying to toots
Add a new in_reply_to ctor argument to ComposeScreen; if specified, this
is a reply.  Take the visibility from the status being replied to, and
pre-set the text area to the username of the original poster.

When replying to a reblogged status, this does not include the username
of the reblogger in the default compose text.  This matches the
behaviour of the Mastodon web UI, and most (but not all) clients.  This
could be a user preference, since people have different opinions on how
this should work.

Add a new 'r' binding in TimelineScreen to compose a reply.

While here, fix a bug that posting a status without a content warning
would fail.