timeline: use server-side default expand preference
Use the server-side preference 'reading:expand:spoilers' to decide
whether to expand content warnings by default.  If not set, default to
false.

Change the -S/--always-show-sensitive flag into a three-way switch; not
set means to use the server-side preference, while setting it to on or
off will override the server preference.