timeline: allow refreshing
Add a new action to refresh the current timeline.  Since Textual doesn't
seem to allow binding to ',', use '.' for this instead.

This requires that the timeline recreate the generator, so introduce a
new api.Timeline class that represents a high-level view of a timeline.
For now this only provides one member, create_generator().  TooiApp now
passes a Timeline to TimelineScreen rather than a generator.

While here, split the 'public' timeline into 'local' and 'federated'
timelines to match other clients.