Add tab support
Initial pass at tab support, using Textual's TabbedContent widget.
TabbedContent is a bit lacking in features, but it seems more sensible
to use the built-in functionality and improve it if necessary rather
than trying to re-invent this.

Each timeline is now displayed as a TimelineTab (nee TimelineScreen) in
the new MainScreen screen.  By default, the home timeline is loaded;
using 'g' to switch to another timeline opens the new timeline in a tab.

Tabs are persistent, so switching between timelines doesn't
automatically reload the timeline, the user needs to press the 'refresh'
key.

Tabs can be navigated using the mouse, the keyboard (when the tab bar is
focussed) or by using the number keys 1..9 with 0=10 to switch to a
specific tab.  Press ^D to close the current tab.

The default Textual tab bar is quite large and uses a lot of space.  I
think it would be better to eventually replace this with a one-line tab
bar integrated into the top status bar.  In the mean time, I've restyled
the status list to make its appearance match the tab bar a bit better.