ComposeScreen: support Hometown's unfederated posts (#13)
This is a Hometown-specific feature which allows posts to be marked as
unfederated.

Add a local_only option to api.statuses.post() to set the local_only
flag in the status.  By default this is None, so it won't be sent to
Mastodon servers.

Detect support for unfederated posts in ComposeScreen by checking if the
posting:default:federation preference is present.  If it is, display a
new post option for selecting the federation status.

If reply to a post with the local_only flag, default to the federation
status of the post being replied to rather than the user preference.