EventList: support relative timestamps
Enabled by the --relative-timestamps option.  Since it's unlikely anyone
will use this option on the command line (as opposed to putting it in
the config file) use an uppercase short option, -R.

If enabled, EventListItem will render the timestamps as the first of
"<X>d", "<X>h", "<X>m" or "<X>s" for which <X> is >= 1.  Only integer
values are rendered.