EventList: limit size of the list
When prepending items to the beginning of the list (which happens when
loading new statuses), limit the size of the list to self.MAX_LENGTH to
prevent it from growing forever.

MAX_LENGTH is currently set to 1024, which should be more than enough
for most users.

This doesn't affect scrolling backward through the timeline: events are
only removed when the timeline is refreshed.