EventList: allow existing events to be updated
Add a refresh_event() method to EventListItem which updates the event.
Currently this is only used to update the timestamp when relative
timestamps are enabled, but in future we can use this to update
favourite/boost status as well.

Add refresh_events() to EventList which calls refresh_event() on all
EventListItems, and call it from refresh_timeline() in TimelineTab.