Don't filter notifications by type
By showing all notifications, even if we don't know how to properly
display them, we'll make it more obvious what we still need to
implement.

BTW, filtering was already not working because the invocation of
notification_generator passed Params as the first parameter instead of
second. I removed the path parameter from that function since it's not
used.