timeline: use a queue for event updates
Change timeline.update() to queue the update to a background async task
and return immediately.  The background task will add new events to the
timeline's internal queue, which can be accessed using the new
get_events() and get_events_nowait() members.

In TimelineTab, create a background async task that calls get_events()
on the timeline forever, and posts NewEventPosted() when a new event
arrives.  Add a handler in the UI task which hands the new events to the
EventList.

This allows Timeline to push unsolicited new events to TimelineTab
without having to wait for user input, which will be used for automatic
timeline refresh and streaming.