timeline: prevent concurrent update()s
This could happen if the user presses '.' multiple times, or if they
refresh while a background update is happening.  This could cause
various issues such as the same event being posted multiple times.

To avoid this, add a new _interlocked_update() method to Timeline that
tracks if an update is already running; if it is, simply return without
doing anything.