EventListItem: display more flags
Display a flag if we favourited or boosted a status.  The flags column
is now 4 characters for every type of status, although currently it
doesn't show boost/favourite flags on notification events.

Both 'we boosted this status' and 'this status was boosted by someone
else' use the B flag, but in different positions, which might be
confusing; perhaps there's a better way to indicate this.