Add support for external image viewer
The image viewer is configured by the media.image_viewer option.

When viewing media, we download all of the status's attachments to a
temporary directory and invoke the image viewer on the downloaded files,
passing each file as an argv element.  This happens in a @work job, so
the image viewer doesn't block the UI.

Tested with swayimg.