api.streaming: implement the protocol correctly
The Mastodon API docs are not very clear about this, but according to
the MDN documentation[0] each event can be multiple lines in length and
will be terminated by a blank line.  Fix _handle_line() to implement the
protocol like this, which fixes occasional issues with event payloads
that had embedded newlines.

[0] https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events