api.timeline: handle duplicate statuses in stream()
If the server pushes an event to the stream, and we call _update()
before we handle it, and the event is also returned by the update, then
it will be added to the timeline twice.  Fix this by tracking events
received by both updates and streaming, and eliding duplicates in either
case.

In theory this could allow self._seen_events to grow forever if _update
is never called, but we forcibly enable timeline_refresh if streaming is
enabled, so that should never happen.